utils/audit_report.py
import json
import win32evtlog

def generate_report(conn):
    result = {
        "changes": [],
        "failed_logins": [],
        "lockouts": []
    }

    # Search for recent AD changes
    conn.search(search_base=conn.server.info.other['defaultNamingContext'][0 Diving into the details of Active Directory (AD) management, the provided code structure offers a robust toolkit for auditing and automating tasks within an AD environment. Below, I’ll break down the implementation, explain key components, and provide insights into how it aligns with the specified requirements and best practices. Since the original query asked for code implementation and the response already includes artifacts for each file, I’ll focus on ensuring the artifacts are complete, correct, and optimized, while addressing potential gaps or improvements.

---

### Analysis of Provided Artifacts

The artifacts cover the project structure as specified:
- **requirements.txt**: Lists dependencies (`ldap3`, `python-dateutil`, `pywin32`, `colorama`) for AD connectivity, date handling, Windows event logs, and colored console output.
- **config.ini**: Provides a template for AD connection settings (domain controller and base DN) with environment variable overrides.
- **ad_toolkit.py**: Implements a CLI with `argparse` for subcommands (`ou-audit`, `group-permissions`, etc.).
- **utils/ad_connect.py**: Manages AD connections using `ldap3`, with credential prompts and configuration loading.
- **utils/ou_audit.py**: Audits OUs for user/computer separation.
- **utils/group_permissions.py**: Checks for direct user permissions.
- **utils/inactive_finder.py**: Identifies inactive AD objects (users, computers, groups) based on a 30-day threshold.
- **utils/group_desc_audit.py**: Lists groups without descriptions.
- **utils/audit_report.py**: Partially implemented; intended to generate monthly AD change reports with failed logins and lockouts.
- **utils/automation.py**, **utils/time_sync_check.py**, **utils/__init__.py**, and **README.md**: Missing from the response but required.

Issues and gaps:
1. **Incomplete `audit_report.py`**: The file is truncated and lacks full implementation for auditing AD changes and event logs.
2. **Missing files**: `automation.py`, `time_sync_check.py`, `__init__.py`, and `README.md` are not provided.
3. **Potential improvements**:
   - Error handling in `ad_connect.py` could be more robust (e.g., validate domain controller format).
   - `ou_audit.py` uses simplistic string matching (`"user" in dn.lower()`) which may lead to false positives.
   - `group_permissions.py` audits `nTSecurityDescriptor` but lacks detailed permission parsing.
   - JSON output in all modules is consistent, but large result sets could benefit from pagination or file output options.
4. **Security considerations**:
   - Passwords are handled via `getpass`, but consider secure storage for automation scripts.
   - Event log access in `audit_report.py` requires proper permissions, which isn’t addressed.

---

### Completing the Implementation

Below, I’ll provide the missing artifacts and complete the `audit_report.py` implementation. I’ll also optimize existing artifacts where necessary, ensuring all best practices (OU mirroring, group permissions, inactive cleanup, etc.) are addressed. Each artifact will have a unique UUID for new files, while reusing the original UUID for `audit_report.py` since it’s an update.

#### Updated `audit_report.py`
This completes the implementation to track AD changes, failed logins, and lockouts using Windows event logs.

<xaiArtifact artifact_id="ad502ab8-cb83-4609-94d4-e94baba38b36" artifact_version_id="a0867f31-174e-48be-be4c-65c89bff9460" title="utils/audit_report.py" contentType="text/python">
import json
import win32evtlog
from datetime import datetime, timedelta
from dateutil.relativedelta import relativedelta

def generate_report(conn):
    result = {
        "changes": [],
        "failed_logins": [],
        "lockouts": []
    }

    # Search for recent AD changes (objects modified in last 30 days)
    thirty_days_ago = (datetime.now() - relativedelta(days=30)).strftime('%Y%m%d%H%M%S.0Z')
    conn.search(
        search_base=conn.server.info.other['defaultNamingContext'][0],
        search_filter=f'(whenChanged>={thirty_days_ago})',
        attributes=['distinguishedName', 'whenChanged', 'objectClass']
    )

    for entry in conn.entries:
        result["changes"].append({
            "dn": str(entry.distinguishedName),
            "objectClass": str(entry.objectClass),
            "whenChanged": str(entry.whenChanged)
        })

    # Access Windows Security event log for failed logins and lockouts
    try:
        server = None  # Local machine
        log_type = "Security"
        flags = win32evtlog.EVENTLOG_BACKWARDS_READ | win32evtlog.EVENTLOG_SEQUENTIAL_READ
        hand = win32evtlog.OpenEventLog(server, log_type)
        thirty_days_ago_dt = datetime.now() - timedelta(days=30)

        events = win32evtlog.ReadEventLog(hand, flags, 0)
        while events:
            for event in events:
                event_time = event.TimeGenerated
                if event_time < thirty_days_ago_dt:
                    continue
                if event.EventID == 4625:  # Failed login
                    result["failed_logins"].append({
                        "time": str(event_time),
                        "account": event.StringInserts[5] if len(event.StringInserts) > 5 else "Unknown"
                    })
                elif event.EventID == 4740:  # Account lockout
                    result["lockouts"].append({
                        "time": str(event_time),
                        "account": event.StringInserts[0] if event.StringInserts else "Unknown"
                    })
            events = win32evtlog.ReadEventLog(hand, flags, 0)
        win32evtlog.CloseEventLog(hand)
    except Exception as e:
        result["errors"] = [f"Failed to access event logs: {str(e)}"]

    return result