requirements.txt
ldap3==2.9.1
python-dateutil==2.8.2
pywin32==306
colorama==0.4.6